requirements.txt
python-dotenv
ipykernel
langchain
langchain_groq
langchain_core
streamlit